Add pc anywhere to gen4
Add the Pc anywhere (or box link whatev you name it) to the R button. Currently crashes in ANY area that does not has a PC in it (so any area that is not a Pokemon Center). It is fixed locally using DS Pokemon ROM Editor, need to find a way to commit the change over.